Usuarios - Py/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose
psycopg2-binary
